requirements.txt
asgiref==3.7.1
backports.zoneinfo==0.2.0
Django==3.2.25
django-cors-headers==4.1.0
django-filter==2.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
gunicorn==22.0.0
packaging==24.1
Pillow==8.0.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.0
typing-extensions==4.7.1
